Bug report added to bottom of README
Submitting this as a PR since I can't raise an issue on the fork. Would be good to have this as its own repo since it does something different to the one it was forked off of.

Occasionally this will randomly happen:

```
var sig = secp256k1.sign(msgHash, privateKey)
                      ^

TypeError: private key should be a Buffer
    at Object.exports.ecsign (.../node_modules/ethereumjs-tx/node_modules/ethereumjs-util/index.js:328:23)
    at Transaction.sign (.../node_modules/ethereumjs-tx/es5/index.js:252:23)
```

But trying again will usually fix it. Not sure why this only happens occasionally.